pyproject.toml
[project]
name = "p3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pydicom>=3.0.1",
    "pymongo>=4.15.3",
    "scikit-image>=0.25.2",
    "seaborn>=0.13.2",
]
[tool.hatch.build.targets.wheel]
packages = ["src/de_p3"]

[project.scripts]
run-pipeline-mongo = "de_p3:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"